Agregado funcionalidad de routes carrito. Falta el ultimo metodo

+++ src/services/products/delByIdService.ts
@@ -1,9 +1,15 @@
-import { deleteById, contenedor } from "../../utils/Container";
+import { deleteById, contenedor } from "../../utils/ContainerProducts";
+import { Request } from 'express';
 
-const delByIdService = async (req: any) => {
+const delByIdService = async (req: Request) => {
   const { id } = req.params;
-  const products = await deleteById(contenedor, parseInt(id));
-  return products;
-};
+  const { admin } = req.body;
+  if(admin == 1){
+    const products = await deleteById(contenedor, parseInt(id));
+    return products;
+  }else {
+    return "Acceso denegado";
+  };
+}
 
 export {delByIdService};
